Refs #16784. Separation of transport and descriptor.
